fix(base): the background worker threads of v8::Platform can not be started when reinitialize v8::Platform
We need the feature that the background worker threads of v8::Platform can be stoped and started many times.

Because of the PHP's process model, we should stop all background threads before the process being forked, and restart the threads in the child processes.

But the v8::V8::InitializePlatform function can only be called once, and the default platform does not provide api to stop threads. So we need to implement a custom platform.

To make it simple, our custom platform just wraps the default platform and forwards corresponing functions. When the threads need to be stoped, the inner default platform instance will be deleted. We should not call any v8 function during the shutdown time.

Additionally, we do not use v8::TracingController and v8::PageAllocator of default platform, in order to avoid null pointer exception.